docs: update README, add doc strings to all exported functionality
