+++ cmd/export/main.go
@@ -1,3 +1,6 @@
+// Small utility to export the HIBP data to stdout.
+// Expects the data to be available in the default data directory or in the directory specified as the first argument.
+// Data is expected to be compressed.
 package main
 
 import (
